ajout creation commande

+++ microservice/client/src/main/java/com/example/client/beans/ProductBean.java
@@ -8,12 +8,12 @@ public class ProductBean {
     private String description;
     private String illustration;
 
-    private double price;
+    private Float price;
 
     public ProductBean() {
     }
 
-    public ProductBean(Long id, String name, String description, String illustration, double price) {
+    public ProductBean(Long id, String name, String description, String illustration, Float price) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -45,11 +45,11 @@ public class ProductBean {
         this.illustration = illustration;
     }
 
-    public double getPrice() {
+    public Float getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(Float price) {
         this.price = price;
     }
